organized code and made css for moves abilites and such match search modal style
