+++ client/src/styles/base.css
@@ -1,0 +1,64 @@
+:root {
+  --color-1: #fff2e0;  /* Background */
+  --color-2: #c0c9ee;  /* Surfaces / cards */
+  --color-3: #a2aadb;  /* Primary accent (buttons, links) */
+  --color-4: #898ac4;  /* Secondary accent (hover, alerts) */
+}
+
+.App {
+  text-align: center;
+  background-color: var(--color-1);
+  width: 100vw;
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+}
+
+.navbar {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 60px;
+  background-color: var(--color-2);
+  color: black;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 0 20px;
+  z-index: 1000;
+  font-size: 1.5rem;
+  font-weight: bold;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.Debug-Bar {
+  position: fixed;
+  bottom: 30px;
+  height: 60px;
+  width: 100%;
+  background-color: var(--color-2);
+  color: black;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 16px;
+  font-size: 1.5rem;
+  font-weight: bold;
+  box-shadow: 0 -2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.Debug-Bar button {
+  padding: 10px 20px;
+  background-color: var(--color-3);
+  color: white;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  z-index: 1500;
+  transition: background-color 0.2s ease;
+}
+
+.Debug-Bar button:hover {
+  background-color: var(--color-4);
+}
